refactor(file-name): renamed file
- renamed clan.ts to clanName.ts to reflect the name of the enum within
